refactor(eslint): relocate node-only common files and bridge renderer import helpers
- private-host, bundle-spectral-ruleset -> main/; send-request -> network/
- import.ts: __IS_RENDERER__ forks via ui/utils/import-bridge (no window token)
- drop all renderer node-builtin exemptions from eslint config
