+++ packages/insomnia-inso/src/cli.ts
@@ -10,7 +10,7 @@ import { Confirm } from 'enquirer';
 import { pick } from 'es-toolkit';
 import { isDevelopment, JSON_ORDER_PREFIX, JSON_ORDER_SEPARATOR } from 'insomnia/src/common/constants';
 import { insomniaFetch } from 'insomnia/src/common/insomnia-fetch';
-import { getSendRequestCallbackMemDb } from 'insomnia/src/common/send-request';
+import { getSendRequestCallbackMemDb } from 'insomnia/src/network/send-request';
 import { initRuntime } from 'insomnia/src/runtimes';
 import { nodeRuntime } from 'insomnia/src/runtimes/runtime.node';
 import { configureFetch } from 'insomnia-api';
